requirements.txt
django==1.11.16
django-ckeditor
Pillow
